[WIP] more refactoring, renaming ..
