--- Classes/class.tx_ttnews_templateeval.php
@@ -1,68 +1,0 @@
-<?php
-namespace RG\TtNews;
-/***************************************************************
- *  Copyright notice
- *
- *  (c) 2009 Dmitry Dulepov (dmitry@typo3.org)
- *  All rights reserved
- *
- *  This script is part of the TYPO3 project. The TYPO3 project is
- *  free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2 of the License, or
- *  (at your option) any later version.
- *
- *  The GNU General Public License can be found at
- *  http://www.gnu.org/copyleft/gpl.html.
- *  A copy is found in the textfile GPL.txt and important notices to the license
- *  from the author is found in LICENSE.txt distributed with these scripts.
- *
- *
- *  This script is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  This copyright notice MUST APPEAR in all copies of the script!
- ***************************************************************/
-/**
- * Hook class for TCEforms.
- *
- * $Id$
- *
- * @author Dmitry Dulepov <dmitry@typo3.org>
- */
-/**
- * [CLASS/FUNCTION INDEX of SCRIPT]
- *
- *
- *
- *   52: class tx_ttnews_templateeval
- *   59:     function deevaluateFieldValue($params)
- *
- * TOTAL FUNCTIONS: 1
- * (This index is automatically created/updated by the extension "extdeveval")
- *
- */
-
-/**
- * Hook class for BE to show nice template name.
- *
- * @author	Dmitry Dulepov
- */
-class tx_ttnews_templateeval {
-	/**
- * Hook to show proper template name in BE if entry was created using older tt_news version
- *
- * @param	array		$params	Params to the hook
- * @return	string		Evaluated value
- */
-	function deevaluateFieldValue($params) {
-		if (trim($params['value']) != '' && false === strpos($params['value'], '/')) {
-			$params['value'] = 'uploads/tt_news/' . $params['value'];
-		}
-		return $params['value'];
-	}
-}
-
-
